Multi-Drive Issue Resolved
In this edition, both Intrface(AXI) and custom Logic can write the slave
registers.

Although in most cases, this feature is not that useful(Designer can
decide which slave register will only be wrote by the Interface, which
will only be wrote by the custom logic. In some other cases, this
feature is useful, for example, the Implementation of Interupt State
Register(ISR) is the one need to be wrote both by the AXI and the
custom logic.